MATCH-V5 retrofit API
0. Add retrofit and LoL services API to matchData dependencies
1. Retrofit interface and methods
2. Match V5 DTOs
3. Match data objects
4. Match repository
5. Set up JUnit 5
6. Create unit tests from match data repository
7. Major dependencies upgrade

+++ findSummonerImpl/src/main/java/com/tsarsprocket/reportmid/findSummonerImpl/viewIntent/InternalViewIntent.kt
@@ -1,7 +1,7 @@
 package com.tsarsprocket.reportmid.findSummonerImpl.viewIntent
 
 import com.tsarsprocket.reportmid.kspApi.annotation.Intent
-import com.tsarsprocket.reportmid.lol.model.Region
+import com.tsarsprocket.reportmid.lol.api.model.Region
 import com.tsarsprocket.reportmid.viewStateApi.viewIntent.ViewIntent
 import kotlinx.parcelize.Parcelize
 
@@ -24,7 +24,7 @@ internal sealed interface InternalViewIntent : ViewIntent {
     @Parcelize
     @Intent
     data class RegionSelected(
-        val newRegionId: Long
+        val newRegionId: Int
     ) : InternalViewIntent
 
     @Parcelize
